Updated project structure to ease SPM changes.
